backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
razorpay
pydantic
requests
